+++ sumologic/resource_sumologic_monitors_library_monitor_test.go
@@ -138,6 +138,7 @@ func TestAccSumologicMonitorsLibraryMonitor_create(t *testing.T) {
 	testContentType := "Monitor"
 	testMonitorType := "Logs"
 	testIsDisabled := false
+	testEvaluationDelay := "5m"
 	testQueries := []MonitorQuery{
 		{
 			RowID: "A",
@@ -203,6 +204,7 @@ func TestAccSumologicMonitorsLibraryMonitor_create(t *testing.T) {
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "name", testName),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "type", testType),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "description", testDescription),
+					resource.TestCheckResourceAttr("sumologic_monitor.test", "evaluation_delay", testEvaluationDelay),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "content_type", testContentType),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "queries.0.row_id", testQueries[0].RowID),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "triggers.0.trigger_type", testTriggers[0].TriggerType),
@@ -250,6 +252,7 @@ func TestAccSumologicMonitorsLibraryMonitor_update(t *testing.T) {
 	testContentType := "Monitor"
 	testMonitorType := "Logs"
 	testIsDisabled := false
+	testEvaluationDelay := "5m"
 	testQueries := []MonitorQuery{
 		{
 			RowID: "A",
@@ -307,6 +310,7 @@ func TestAccSumologicMonitorsLibraryMonitor_update(t *testing.T) {
 	testUpdatedContentType := "Monitor"
 	testUpdatedMonitorType := "Logs"
 	testUpdatedIsDisabled := true
+	testUpdatedEvaluationDelay := "8m"
 	testUpdatedQueries := []MonitorQuery{
 		{
 			RowID: "A",
@@ -372,6 +376,7 @@ func TestAccSumologicMonitorsLibraryMonitor_update(t *testing.T) {
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "name", testName),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "type", testType),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "description", testDescription),
+					resource.TestCheckResourceAttr("sumologic_monitor.test", "evaluation_delay", testEvaluationDelay),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "content_type", testContentType),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "queries.0.row_id", testQueries[0].RowID),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "triggers.0.trigger_type", testTriggers[0].TriggerType),
@@ -387,6 +392,7 @@ func TestAccSumologicMonitorsLibraryMonitor_update(t *testing.T) {
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "name", testUpdatedName),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "type", testUpdatedType),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "description", testUpdatedDescription),
+					resource.TestCheckResourceAttr("sumologic_monitor.test", "evaluation_delay", testUpdatedEvaluationDelay),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "content_type", testUpdatedContentType),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "queries.0.row_id", testUpdatedQueries[0].RowID),
 					resource.TestCheckResourceAttr("sumologic_monitor.test", "triggers.0.trigger_type", testUpdatedTriggers[0].TriggerType),
@@ -446,6 +452,7 @@ func testAccCheckMonitorsLibraryMonitorAttributes(name string) resource.TestChec
 			resource.TestCheckResourceAttrSet(name, "created_by"),
 			resource.TestCheckResourceAttrSet(name, "is_locked"),
 			resource.TestCheckResourceAttrSet(name, "monitor_type"),
+			resource.TestCheckResourceAttrSet(name, "evaluation_delay"),
 			resource.TestCheckResourceAttrSet(name, "is_system"),
 			resource.TestCheckResourceAttrSet(name, "is_disabled"),
 			resource.TestCheckResourceAttrSet(name, "name"),
@@ -471,6 +478,7 @@ resource "sumologic_monitor" "test" {
 	is_disabled = false
 	content_type = "Monitor"
 	monitor_type = "Logs"
+	evaluation_delay = "5m"
 	queries {
 		row_id = "A"
 		query = "_sourceCategory=monitor-manager error"
@@ -515,6 +523,7 @@ resource "sumologic_monitor" "test" {
 	is_disabled = true
 	content_type = "Monitor"
 	monitor_type = "Logs"
+	evaluation_delay = "8m"
 	queries {
 		row_id = "A"
 		query = "_sourceCategory=monitor-manager info"