+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-## 2.9.10 (Unreleased)
+## 2.9.11 (Unreleased)
+
+
+## 2.9.10 (August 24, 2021)
 
 FEATURES: